join_notification/requirements.txt
watchfiles==1.0.4
discord.py==2.5.2